% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EM}
\alias{EM}
\title{One EM iterate}
\usage{
EM(oldpara, data, lambda)
}
\arguments{
\item{oldpara}{parameter set from the last step}

\item{data}{postprocessed data}

\item{lambda}{tuning parameter}
}
\value{
new parameter set
}
\description{
One EM iterate
}
\keyword{internal}
