% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gen_data}
\alias{gen_data}
\title{Generate simulation data}
\usage{
gen_data(n, varres, gama = 2, type = "n", ka = 2L, kb = 2L)
}
\arguments{
\item{n}{number of subjects}

\item{varres}{variance or conditional variance of residuals}

\item{gama}{the degree of freedoms for slash and t distribution}

\item{type}{model type. 'n' means normal, 't' means t distribution, 's' means slash distribution}

\item{ka}{number of pcs for the first response variable (Y)}

\item{kb}{number of pcs for the second response variable (Z)}
}
\value{
simulated data
}
\description{
Use this function to generate data to get familiar with this package. The simulated data have the same setting as the paper describes.
}
\examples{
rawdata = gen_data(n=100,varres=0.01, gama=2, type='t',ka=2,kb=2)
}
