% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kfold_CV.R
\name{Kf_CV}
\alias{Kf_CV}
\title{K-fold Cross-Validation}
\usage{
Kf_CV(data, lambda, K, ka, kb)
}
\arguments{
\item{data}{precessed data}

\item{lambda}{the tuning parameter}

\item{K}{number of folds}

\item{ka}{number of pcs for the first response variable}

\item{kb}{number of pcs for the second response variable}
}
\value{
the sum of mean square error of K folds
}
\description{
Divide the whole dataset into K folds. Take K-1 folds as training dataset 
and take the rest one as the testing dataset. Return mean square error (MSE).
}
\examples{
rawdata = gen_data(n=50)
visit = seq(0,100,20)
data = predata(nobs = rawdata$nobs, 
               time = rawdata$time, 
               y = rawdata$y, 
               z = rawdata$z, 
               knots = visit, 
               order = 3)
MSE = matrix(nrow=5,ncol=5)
for(i in 1:5)
  for(j in 1:5)
  {
     lambda = c(200*i, 200*i, 4000*j, 4000*j)
     MSE[i,j] = Kf_CV(data, lambda, K=10, ka=1,kb=2)
  }

}
