% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglike}
\alias{loglike}
\title{Evaluate the loglikelihood value}
\usage{
loglike(data, para, type)
}
\arguments{
\item{data}{postprocessed data}

\item{para}{parameter set}
}
\value{
-2*log-likelihood value
}
\description{
Evaluate the loglikelihood value
}
\keyword{internal}
