% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSEcirteria.R
\name{MSE}
\alias{MSE}
\title{Evaluate the MSE}
\usage{
MSE(data, para)
}
\arguments{
\item{data}{preprocessed data}

\item{para}{parameter set}
}
\value{
MSE
}
\description{
Evaluate the MSE
}
\keyword{internal}
