% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGen.R
\name{mu_t}
\alias{mu_t}
\title{Simulation function mu(t)
mean function of the first functional variable}
\usage{
mu_t(t)
}
\arguments{
\item{t}{covariate}
}
\value{
mean function mu(t)
}
\description{
Simulation function mu(t)
mean function of the first functional variable
}
