% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGen.R
\name{nu_t}
\alias{nu_t}
\title{Simulation function nu(t)
mean function of the second functional variable}
\usage{
nu_t(t)
}
\arguments{
\item{t}{covariate}
}
\value{
mean function nu(t)
}
\description{
Simulation function nu(t)
mean function of the second functional variable
}
