% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{orth_algo}
\alias{orth_algo}
\title{orthogonalize the covariance matrix}
\usage{
orth_algo(Th, V)
}
\arguments{
\item{Th}{theta before orthogonalize}

\item{V}{covariance matrix before}
}
\value{
orthogonalized theta and covariance matrix
}
\description{
orthogonalize the covariance matrix
}
\keyword{internal}
