% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{orthbasis}
\alias{orthbasis}
\title{Create bsplines}
\usage{
orthbasis(x, knots = 10, order = 3)
}
\arguments{
\item{x}{The points you are interested in.}

\item{knots}{The full set of knots used to define the basis functions. It can be a number or a vector. If it is a number, the full set of knots is seq(0,1,1/knots); if it is a vector, knots is this vector.}

\item{order}{Order of bsplines. Default is 3 which means cubic splines.}
}
\value{
A list including three elements: The full set of knots (knots), design matrix (B) and penalty matrix (Omega).
}
\description{
Create bsplines
}
\details{
Suppose we create n othogonal bspline basis b=(b_1,...,b_n), where b_i is a function. 
Then the design matrix B = b(x) = (b_1(x),...,b_n(x)). 
The number of rows is equal to the length of x. The number of columns is equal to the number of basis.
The penalty matrix Omega = integral b''(t)b''(t)^T dt.
}
\examples{
rawdata = gen_data(n=100,varres=0.01, gama=2, type='t',ka=2,kb=2)
time = rawdata$dataset[,2]
spl_info = orthbasis(time,knots = 8, order =3)
}
