% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt.data.R
\name{plt.data}
\alias{plt.data}
\title{Plot the raw data}
\usage{
plt.data(data)
}
\arguments{
\item{data}{Processed data. Use "predata" to preprocess the raw data first.}
}
\description{
display two plots of the paired raw data
}
\examples{
rawdata = gen_data(n=100,varres=0.01, gama=2, type='t',ka=2,kb=2)
data = predata(nobs_y = rawdata$obs_times, 
               nobs_z = rawdata$obs_times, 
               time_y = rawdata$dataset[,2], 
               time_z = rawdata$dataset[,2], 
               y = rawdata$dataset[,3], 
               z = rawdata$dataset[,4], 
               knots = 10, 
               order=3)
plt.data(data)
}
