% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predata.R
\name{predata}
\alias{predata}
\title{Preprocessing data}
\usage{
predata(nobs_y, nobs_z, time_y, time_z, y, z, knots = 10, order = 3)
}
\arguments{
\item{nobs_y}{A vector whose component is the number of Y observations of each subject}

\item{nobs_z}{A vector whose component is the number of Z observations of each subject}

\item{time_y}{A vector records Y observation time of all subjects}

\item{time_z}{A vector records Z observation time of all subjects}

\item{y}{A vector records first response variable of all subjects}

\item{z}{A vector records second response variable of all subjects}

\item{knots}{The full set of knots used to define the basis functions. It can be a number or a vector. If it is a number, the full set of knots is  seq(0,1,1/knots); if it is a vector, the full set of knots is this vector.}

\item{order}{Order of bsplines}
}
\value{
A list including number of subjects (n), obs times (nobs_y and nobs_z), the time we observe data (time_y and time_z),paired data (y and z), the full set of knots (knots), design matrix (B_y and B_z) and penalty matrix (Omega)
}
\description{
Convert the raw data to a data structure that can be used in this package. The new data structure includes two important entries, the design matrix and the penalty matrix.
}
\details{
Suppose the data include the information of n subjects.

For each subject, there are m_i^{(y)} (i=1,...n) observations of Y which are observed at time \eqn{t_i =  (t_{i1}^{(y)},...t_{i m_i}^{(y)})}.

For each subject, there are m_i^{(z)} (i=1,...n) observations of Z which are observed at time \eqn{t_i =  (t_{i1}^{(z)},...t_{i m_i}^{(z)})}.

Suppose the data are generated from potential curves Y_i and Z_i. Then

nobs_y = (m_1^{(y)},...,m_n^{(y)}), time_y = (t_1^{(y)},...t_n^{(y)}), y = (Y_1(t_1^{(y)}),...,Y_n(t_n^{(y)})),

nobs_z = (m_1^{(z)},...,m_n^{(z)}), time_z = (t_1^{(z)},...t_n^{(z)}), z = (Z_1(t_1^{(z)}),...,Z_n(t_n^{(z)})).
}
\examples{
n = 5
nobs = 1+rbinom(n,5,0.9)
time=c()
for(i in 1:n)
{
  time=c(time,0,runif(nobs[i]-1))
}
sumobs = sum(nobs)
y = rnorm(sumobs,0,1.5)
z = rnorm(sumobs,1,1)
data = predata(nobs,time,y,z,knots = 8,order=3)
rawdata = gen_data(n=100,varres=0.01, gama=2, type='t',ka=2,kb=2)
data = predata(nobs_y = rawdata$obs_times, 
               nobs_z = rawdata$obs_times, 
               time_y = rawdata$dataset[,2], 
               time_z = rawdata$dataset[,2], 
               y = rawdata$dataset[,3], 
               z = rawdata$dataset[,4], 
               knots = 10, 
               order=3)
}
\seealso{
\code{\link{orthbasis}}
}
