% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGen.R
\name{gen_data}
\alias{gen_data}
\title{Generate Data}
\usage{
gen_data(n)
}
\arguments{
\item{n}{number of subjects}
}
\value{
a list of 4 components including nobs, time, y and z
}
\description{
Generate the simulation data described in the paper.
}
\examples{
rawdata = gen_data(50)
}
