% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{orthbasis}
\alias{orthbasis}
\title{Create bsplines}
\usage{
orthbasis(x, knots = 8, order = 3)
}
\arguments{
\item{x}{the points you are interested in}

\item{knots}{if it is a number, the knots = seq(0,1,1/nknots); if it is a vector, knots is this vector}

\item{order}{default is 3, it means cubic spline}
}
\value{
a list including two elements. B(x) and covariance matrix of bsplines
}
\description{
Create bsplines
}
