% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt.data.R
\name{plt.data}
\alias{plt.data}
\title{Plot the raw data}
\usage{
plt.data(data)
}
\arguments{
\item{data}{Processed data. Use "predata" to preprocess the raw data first.}
}
\description{
display two plots of the paired raw data
}
\examples{
plt.data(data)
}
