% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predata.R
\name{predata}
\alias{predata}
\title{Preprocessing data}
\usage{
predata(nobs, time, y, z, knots = 8, order = 3)
}
\arguments{
\item{nobs}{A vector whose component is the number of the observations of each subject}

\item{time}{A vector records the observation time of all subjects}

\item{y}{A vector records first response variable of all subjects}

\item{z}{A vector records second response variable of all subjects}

\item{knots}{The full set of knots used to define the basis functions. It can be a number or a vector. If it is a number, the knots = seq(0,1,1/knots); if it is a vector, knots is this vector.}

\item{order}{Order of bsplines}
}
\value{
A list including number of subjects (n), obs times (nobs), paired data (dataset), the full set of knots (knots), design matrix (B) and penalty matrix (Omega)
}
\description{
Change the raw data to a sturcture that can be used in this package. Create the design matrix and penalty matrix.
}
\details{
Suppose the data include the information of n subjects.

For each subject, there are m_i (i=1,...n) observations which are observed at time \eqn{t_i =  (t_{i1},...t_{i m_i})}.

Suppose the data are generated from potential curves Y_i and Z_i. Then

nobs = (m_1,...,m_n), time = (t_1,...t_n), y = (Y_1(t_1),...,Y_n(t_n)), z = (Z_1(t_1),...,Z_n(t_n)).
}
\examples{
n = 5
nobs = 1+rbinom(n,5,0.9)
time=c()
for(i in 1:n)
{
  time=c(time,0,runif(nobs[i]-1))
}
sumobs = sum(nobs)
y = rnorm(sumobs,0,1.5)
z = rnorm(sumobs,1,1)
data = predata(nobs,time,y,z,knots = 8,order=3)

 
rawdata = gen_data(n=50)
visit = seq(0,100,20)
data = predata(nobs = rawdata$nobs, 
               time = rawdata$time, 
               y = rawdata$y, 
               z = rawdata$z, 
               knots = visit, 
               order = 3)
}
